requirements.txt
numpy==1.18.0
tensorboardX==2.0
pyaml==20.4
scipy==1.4
sklearn
pandas==1.0.3
matplotlib==3.2
torchsummaryX==1.1.0
terminalplot==0.3
flask==1.1.2
gunicorn==20.0.4
xgboost==1.2.1
lightgbm==3.0.0
